🏗️(tests) separate team accesses tests by action
Small commit to separate team accesses tests into diferent files.
